Move seearch code from project.rs -> search.rs
