+++ crates/project/src/search.rs
@@ -1,17 +1,34 @@
 use aho_corasick::{AhoCorasick, AhoCorasickBuilder};
-use anyhow::Result;
+use anyhow::{Context, Result};
 use client::proto;
-use language::{char_kind, BufferSnapshot};
+use collections::HashMap;
+use fs::Fs;
+use futures::StreamExt;
+use gpui::{AppContext, BackgroundExecutor, Model, Task};
+use language::{char_kind, Buffer, BufferSnapshot};
 use regex::{Captures, Regex, RegexBuilder};
-use smol::future::yield_now;
+use smol::{
+    channel::{Receiver, Sender},
+    future::yield_now,
+    lock::Semaphore,
+};
 use std::{
     borrow::Cow,
+    cmp::{self, Ordering},
+    collections::VecDeque,
     io::{BufRead, BufReader, Read},
     ops::Range,
-    path::Path,
+    path::{Path, PathBuf},
     sync::{Arc, OnceLock},
 };
-use util::paths::PathMatcher;
+use text::Anchor;
+use util::{
+    paths::{compare_paths, PathMatcher},
+    ResultExt,
+};
+use worktree::{Entry, Snapshot, WorktreeId, WorktreeSettings};
+
+use crate::{buffer_store::BufferStore, worktree_store::WorktreeStore, Item, ProjectPath};
 
 static TEXT_REPLACEMENT_SPECIAL_CHARACTERS_REGEX: OnceLock<Regex> = OnceLock::new();
 
@@ -458,4 +475,519 @@ mod tests {
             }
         }
     }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+enum SearchMatchCandidate {
+    OpenBuffer {
+        buffer: Model<Buffer>,
+        // This might be an unnamed file without representation on filesystem
+        path: Option<Arc<Path>>,
+    },
+    Path {
+        worktree_id: WorktreeId,
+        is_ignored: bool,
+        is_file: bool,
+        path: Arc<Path>,
+    },
+}
+
+pub enum SearchResult {
+    Buffer {
+        buffer: Model<Buffer>,
+        ranges: Vec<Range<Anchor>>,
+    },
+    LimitReached,
+}
+
+pub fn search(
+    buffer_store: &Model<BufferStore>,
+    worktree_store: &Model<WorktreeStore>,
+    fs: Arc<dyn Fs>,
+    query: SearchQuery,
+    cx: &mut AppContext,
+) -> Receiver<SearchResult> {
+    let snapshots = worktree_store
+        .read(cx)
+        .visible_worktrees(cx)
+        .filter_map(|tree| {
+            let tree = tree.read(cx);
+            Some((tree.snapshot(), tree.as_local()?.settings()))
+        })
+        .collect::<Vec<_>>();
+    let include_root = snapshots.len() > 1;
+
+    let mut unnamed_buffers = vec![];
+    let opened_buffers = buffer_store.update(cx, |buffer_store, cx| {
+        buffer_store
+            .buffers()
+            .filter_map(|buffer| {
+                let (is_ignored, snapshot) = buffer.update(cx, |buffer, cx| {
+                    let worktree_store = worktree_store.read(cx);
+                    let is_ignored = buffer
+                        .entry_id(cx)
+                        .and_then(|entry_id| worktree_store.entry_for_id(entry_id, cx))
+                        .map_or(false, |entry| entry.is_ignored);
+                    (is_ignored, buffer.snapshot())
+                });
+                if is_ignored && !query.include_ignored() {
+                    return None;
+                } else if let Some(file) = snapshot.file() {
+                    let matched_path = if include_root {
+                        query.file_matches(Some(&file.full_path(cx)))
+                    } else {
+                        query.file_matches(Some(file.path()))
+                    };
+
+                    if matched_path {
+                        Some((file.path().clone(), (buffer, snapshot)))
+                    } else {
+                        None
+                    }
+                } else {
+                    unnamed_buffers.push(buffer);
+                    None
+                }
+            })
+            .collect()
+    });
+
+    let path_count: usize = snapshots
+        .iter()
+        .map(|(snapshot, _)| {
+            if query.include_ignored() {
+                snapshot.file_count()
+            } else {
+                snapshot.visible_file_count()
+            }
+        })
+        .sum();
+    if path_count == 0 {
+        let (_, rx) = smol::channel::bounded(1024);
+        return rx;
+    }
+    let workers = cx.background_executor().num_cpus().min(path_count);
+    let (matching_paths_tx, matching_paths_rx) = smol::channel::bounded(1024);
+
+    cx.background_executor()
+        .spawn(find_match_candidates(
+            unnamed_buffers,
+            opened_buffers,
+            cx.background_executor().clone(),
+            fs,
+            workers,
+            query.clone(),
+            include_root,
+            path_count,
+            snapshots,
+            matching_paths_tx,
+        ))
+        .detach();
+
+    let (result_tx, result_rx) = smol::channel::bounded(1024);
+    let query = Arc::new(query);
+    let buffer_store = buffer_store.clone();
+
+    cx.spawn(|mut cx| async move {
+        const MAX_SEARCH_RESULT_FILES: usize = 5_000;
+        const MAX_SEARCH_RESULT_RANGES: usize = 10_000;
+
+        let mut matching_paths = matching_paths_rx
+            .take(MAX_SEARCH_RESULT_FILES + 1)
+            .collect::<Vec<_>>()
+            .await;
+        let mut limit_reached = if matching_paths.len() > MAX_SEARCH_RESULT_FILES {
+            matching_paths.pop();
+            true
+        } else {
+            false
+        };
+        cx.update(|cx| {
+            sort_search_matches(&mut matching_paths, cx);
+        })?;
+
+        let mut range_count = 0;
+
+        // Now that we know what paths match the query, we will load at most
+        // 64 buffers at a time to avoid overwhelming the main thread. For each
+        // opened buffer, we will spawn a background task that retrieves all the
+        // ranges in the buffer matched by the query.
+        'outer: for matching_paths_chunk in matching_paths.chunks(64) {
+            let mut chunk_results = Vec::new();
+            for matching_path in matching_paths_chunk {
+                let query = query.clone();
+                let buffer = match matching_path {
+                    SearchMatchCandidate::OpenBuffer { buffer, .. } => {
+                        Task::ready(Ok(buffer.clone()))
+                    }
+                    SearchMatchCandidate::Path {
+                        worktree_id, path, ..
+                    } => buffer_store.update(&mut cx, |buffer_store, cx| {
+                        buffer_store.open_buffer(
+                            ProjectPath {
+                                worktree_id: *worktree_id,
+                                path: path.clone(),
+                            },
+                            cx,
+                        )
+                    })?,
+                };
+
+                chunk_results.push(cx.spawn(|cx| async move {
+                    let buffer = buffer.await?;
+                    let snapshot = buffer.read_with(&cx, |buffer, _| buffer.snapshot())?;
+                    let ranges = cx
+                        .background_executor()
+                        .spawn(async move {
+                            query
+                                .search(&snapshot, None)
+                                .await
+                                .iter()
+                                .map(|range| {
+                                    snapshot.anchor_before(range.start)
+                                        ..snapshot.anchor_after(range.end)
+                                })
+                                .collect::<Vec<_>>()
+                        })
+                        .await;
+                    anyhow::Ok((buffer, ranges))
+                }));
+            }
+
+            let chunk_results = futures::future::join_all(chunk_results).await;
+            for result in chunk_results {
+                if let Some((buffer, ranges)) = result.log_err() {
+                    range_count += ranges.len();
+                    result_tx
+                        .send(SearchResult::Buffer { buffer, ranges })
+                        .await?;
+                    if range_count > MAX_SEARCH_RESULT_RANGES {
+                        limit_reached = true;
+                        break 'outer;
+                    }
+                }
+            }
+        }
+
+        if limit_reached {
+            result_tx.send(SearchResult::LimitReached).await?;
+        }
+
+        anyhow::Ok(())
+    })
+    .detach();
+
+    result_rx
+}
+
+/// Pick paths that might potentially contain a match of a given search query.
+#[allow(clippy::too_many_arguments)]
+async fn find_match_candidates(
+    unnamed_buffers: Vec<Model<Buffer>>,
+    opened_buffers: HashMap<Arc<Path>, (Model<Buffer>, BufferSnapshot)>,
+    executor: BackgroundExecutor,
+    fs: Arc<dyn Fs>,
+    workers: usize,
+    query: SearchQuery,
+    include_root: bool,
+    path_count: usize,
+    snapshots: Vec<(Snapshot, WorktreeSettings)>,
+    matching_paths_tx: Sender<SearchMatchCandidate>,
+) {
+    let fs = &fs;
+    let query = &query;
+    let matching_paths_tx = &matching_paths_tx;
+    let snapshots = &snapshots;
+    for buffer in unnamed_buffers {
+        matching_paths_tx
+            .send(SearchMatchCandidate::OpenBuffer {
+                buffer: buffer.clone(),
+                path: None,
+            })
+            .await
+            .log_err();
+    }
+    for (path, (buffer, _)) in opened_buffers.iter() {
+        matching_paths_tx
+            .send(SearchMatchCandidate::OpenBuffer {
+                buffer: buffer.clone(),
+                path: Some(path.clone()),
+            })
+            .await
+            .log_err();
+    }
+
+    let paths_per_worker = (path_count + workers - 1) / workers;
+
+    executor
+        .scoped(|scope| {
+            let max_concurrent_workers = Arc::new(Semaphore::new(workers));
+
+            for worker_ix in 0..workers {
+                let worker_start_ix = worker_ix * paths_per_worker;
+                let worker_end_ix = worker_start_ix + paths_per_worker;
+                let opened_buffers = opened_buffers.clone();
+                let limiter = Arc::clone(&max_concurrent_workers);
+                scope.spawn({
+                    async move {
+                        let _guard = limiter.acquire().await;
+                        search_snapshots(
+                            snapshots,
+                            worker_start_ix,
+                            worker_end_ix,
+                            query,
+                            matching_paths_tx,
+                            &opened_buffers,
+                            include_root,
+                            fs,
+                        )
+                        .await;
+                    }
+                });
+            }
+
+            if query.include_ignored() {
+                for (snapshot, settings) in snapshots {
+                    for ignored_entry in snapshot.entries(true, 0).filter(|e| e.is_ignored) {
+                        let limiter = Arc::clone(&max_concurrent_workers);
+                        scope.spawn(async move {
+                            let _guard = limiter.acquire().await;
+                            search_ignored_entry(
+                                snapshot,
+                                settings,
+                                ignored_entry,
+                                fs,
+                                query,
+                                matching_paths_tx,
+                            )
+                            .await;
+                        });
+                    }
+                }
+            }
+        })
+        .await;
+}
+
+async fn search_ignored_entry(
+    snapshot: &Snapshot,
+    settings: &WorktreeSettings,
+    ignored_entry: &Entry,
+    fs: &Arc<dyn Fs>,
+    query: &SearchQuery,
+    counter_tx: &Sender<SearchMatchCandidate>,
+) {
+    let mut ignored_paths_to_process =
+        VecDeque::from([snapshot.abs_path().join(&ignored_entry.path)]);
+
+    while let Some(ignored_abs_path) = ignored_paths_to_process.pop_front() {
+        let metadata = fs
+            .metadata(&ignored_abs_path)
+            .await
+            .with_context(|| format!("fetching fs metadata for {ignored_abs_path:?}"))
+            .log_err()
+            .flatten();
+
+        if let Some(fs_metadata) = metadata {
+            if fs_metadata.is_dir {
+                let files = fs
+                    .read_dir(&ignored_abs_path)
+                    .await
+                    .with_context(|| format!("listing ignored path {ignored_abs_path:?}"))
+                    .log_err();
+
+                if let Some(mut subfiles) = files {
+                    while let Some(subfile) = subfiles.next().await {
+                        if let Some(subfile) = subfile.log_err() {
+                            ignored_paths_to_process.push_back(subfile);
+                        }
+                    }
+                }
+            } else if !fs_metadata.is_symlink {
+                if !query.file_matches(Some(&ignored_abs_path))
+                    || settings.is_path_excluded(&ignored_entry.path)
+                {
+                    continue;
+                }
+                let matches = if let Some(file) = fs
+                    .open_sync(&ignored_abs_path)
+                    .await
+                    .with_context(|| format!("Opening ignored path {ignored_abs_path:?}"))
+                    .log_err()
+                {
+                    query.detect(file).unwrap_or(false)
+                } else {
+                    false
+                };
+
+                if matches {
+                    let project_path = SearchMatchCandidate::Path {
+                        worktree_id: snapshot.id(),
+                        path: Arc::from(
+                            ignored_abs_path
+                                .strip_prefix(snapshot.abs_path())
+                                .expect("scanning worktree-related files"),
+                        ),
+                        is_ignored: true,
+                        is_file: ignored_entry.is_file(),
+                    };
+                    if counter_tx.send(project_path).await.is_err() {
+                        return;
+                    }
+                }
+            }
+        }
+    }
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn search_snapshots(
+    snapshots: &Vec<(Snapshot, WorktreeSettings)>,
+    worker_start_ix: usize,
+    worker_end_ix: usize,
+    query: &SearchQuery,
+    results_tx: &Sender<SearchMatchCandidate>,
+    opened_buffers: &HashMap<Arc<Path>, (Model<Buffer>, BufferSnapshot)>,
+    include_root: bool,
+    fs: &Arc<dyn Fs>,
+) {
+    let mut snapshot_start_ix = 0;
+    let mut abs_path = PathBuf::new();
+
+    for (snapshot, _) in snapshots {
+        let snapshot_end_ix = snapshot_start_ix
+            + if query.include_ignored() {
+                snapshot.file_count()
+            } else {
+                snapshot.visible_file_count()
+            };
+        if worker_end_ix <= snapshot_start_ix {
+            break;
+        } else if worker_start_ix > snapshot_end_ix {
+            snapshot_start_ix = snapshot_end_ix;
+            continue;
+        } else {
+            let start_in_snapshot = worker_start_ix.saturating_sub(snapshot_start_ix);
+            let end_in_snapshot = cmp::min(worker_end_ix, snapshot_end_ix) - snapshot_start_ix;
+
+            for entry in snapshot
+                .files(false, start_in_snapshot)
+                .take(end_in_snapshot - start_in_snapshot)
+            {
+                if results_tx.is_closed() {
+                    break;
+                }
+                if opened_buffers.contains_key(&entry.path) {
+                    continue;
+                }
+
+                let matched_path = if include_root {
+                    let mut full_path = PathBuf::from(snapshot.root_name());
+                    full_path.push(&entry.path);
+                    query.file_matches(Some(&full_path))
+                } else {
+                    query.file_matches(Some(&entry.path))
+                };
+
+                let matches = if matched_path {
+                    abs_path.clear();
+                    abs_path.push(&snapshot.abs_path());
+                    abs_path.push(&entry.path);
+                    if let Some(file) = fs.open_sync(&abs_path).await.log_err() {
+                        query.detect(file).unwrap_or(false)
+                    } else {
+                        false
+                    }
+                } else {
+                    false
+                };
+
+                if matches {
+                    let project_path = SearchMatchCandidate::Path {
+                        worktree_id: snapshot.id(),
+                        path: entry.path.clone(),
+                        is_ignored: entry.is_ignored,
+                        is_file: entry.is_file(),
+                    };
+                    if results_tx.send(project_path).await.is_err() {
+                        return;
+                    }
+                }
+            }
+
+            snapshot_start_ix = snapshot_end_ix;
+        }
+    }
+}
+
+fn sort_search_matches(search_matches: &mut Vec<SearchMatchCandidate>, cx: &AppContext) {
+    search_matches.sort_by(|entry_a, entry_b| match (entry_a, entry_b) {
+        (
+            SearchMatchCandidate::OpenBuffer {
+                buffer: buffer_a,
+                path: None,
+            },
+            SearchMatchCandidate::OpenBuffer {
+                buffer: buffer_b,
+                path: None,
+            },
+        ) => buffer_a
+            .read(cx)
+            .remote_id()
+            .cmp(&buffer_b.read(cx).remote_id()),
+        (
+            SearchMatchCandidate::OpenBuffer { path: None, .. },
+            SearchMatchCandidate::Path { .. }
+            | SearchMatchCandidate::OpenBuffer { path: Some(_), .. },
+        ) => Ordering::Less,
+        (
+            SearchMatchCandidate::OpenBuffer { path: Some(_), .. }
+            | SearchMatchCandidate::Path { .. },
+            SearchMatchCandidate::OpenBuffer { path: None, .. },
+        ) => Ordering::Greater,
+        (
+            SearchMatchCandidate::OpenBuffer {
+                path: Some(path_a), ..
+            },
+            SearchMatchCandidate::Path {
+                is_file: is_file_b,
+                path: path_b,
+                ..
+            },
+        ) => compare_paths((path_a.as_ref(), true), (path_b.as_ref(), *is_file_b)),
+        (
+            SearchMatchCandidate::Path {
+                is_file: is_file_a,
+                path: path_a,
+                ..
+            },
+            SearchMatchCandidate::OpenBuffer {
+                path: Some(path_b), ..
+            },
+        ) => compare_paths((path_a.as_ref(), *is_file_a), (path_b.as_ref(), true)),
+        (
+            SearchMatchCandidate::OpenBuffer {
+                path: Some(path_a), ..
+            },
+            SearchMatchCandidate::OpenBuffer {
+                path: Some(path_b), ..
+            },
+        ) => compare_paths((path_a.as_ref(), true), (path_b.as_ref(), true)),
+        (
+            SearchMatchCandidate::Path {
+                worktree_id: worktree_id_a,
+                is_file: is_file_a,
+                path: path_a,
+                ..
+            },
+            SearchMatchCandidate::Path {
+                worktree_id: worktree_id_b,
+                is_file: is_file_b,
+                path: path_b,
+                ..
+            },
+        ) => worktree_id_a.cmp(&worktree_id_b).then_with(|| {
+            compare_paths((path_a.as_ref(), *is_file_a), (path_b.as_ref(), *is_file_b))
+        }),
+    });
 }
